==> 2.Classification.ipynb <==
import pandas as pd

df = pd.read_csv('Data/generated_data.csv')

# Тестирование in-context learning для Mistral-7B-Instruct-v0.2

from vllm import LLM, SamplingParams
from tqdm import tqdm

llm = LLM(model="mistralai/Mistral-7B-Instruct-v0.2")

## Zero-shot learning

prediction = []
sampling_params = SamplingParams(temperature=0, max_tokens=30, stop=['отель', 'самолет'], include_stop_str_in_output=True)
for i in tqdm(range(2000)):
    sample = df.iloc[i]
    prompt = f"""Ты должен отнести текст к одному из двух классов: самолет и отель.
    
    Используй следующий шаблон:
    
    Текст: сообщение пользователя в чате
    Класс: одно из двух значений: самолет или отель
    
    Начинай!
    
    Текст: {sample.text}
    Класс: """
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text.replace(' ','').replace('\n','')
    prediction.append(generated_text)

df['prediction'] = prediction
booking_errors = df[(df['prediction']!=df['label'])&(df['label']=='отель')].shape[0]
ticket_errors = df[(df['prediction']!=df['label'])&(df['label']=='самолет')].shape[0]
print(f'Accuracy booking = {1-booking_errors/1000:.3f}\nAccuracy tickets = {1-ticket_errors/1000:.3f}')

## One-shot learning

prediction_one_shot = []
sampling_params = SamplingParams(temperature=0, max_tokens=30, stop=['отель', 'самолет'], include_stop_str_in_output=True)
for i in tqdm(range(2000)):
    sample = df.iloc[i]
    prompt = f"""Ты должен отнести текст к одному из двух классов: самолет и отель.

    Пример для класса самолет: {df.iloc[0].text}
    Пример для класса отель: {df.iloc[1000].text}
    
    Используй следующий шаблон:
    
    Текст: сообщение пользователя в чате
    Класс: одно из двух значений: самолет или отель
    
    Начинай!
    
    Текст: {sample.text}
    Класс: """
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text.replace(' ','').replace('\n','')
    prediction.append(generated_text)
    prediction_one_shot.append(generated_text)

df['prediction_one_shot'] = prediction_one_shot
booking_errors = df[(df['prediction_one_shot']!=df['label'])&(df['label']=='отель')].shape[0]
ticket_errors = df[(df['prediction_one_shot']!=df['label'])&(df['label']=='самолет')].shape[0]
print(f'Accuracy booking = {1-booking_errors/1000:.3f}\nAccuracy tickets = {1-ticket_errors/1000:.3f}')

# Тестирование in-context learning для Mistral-7B v0.1
(перезапустите ядро)

import pandas as pd

df = pd.read_csv('Data/generated_data.csv')

from vllm import LLM, SamplingParams
from tqdm import tqdm

llm = LLM(model="mistralai/Mistral-7B-v0.1")

## Zero-shot learning

prediction = []
sampling_params = SamplingParams(temperature=0, max_tokens=30, stop=['отель', 'самолет'], include_stop_str_in_output=True)
for i in tqdm(range(2000)):
    sample = df.iloc[i]
    prompt = f"""Ты должен отнести текст к одному из двух классов: самолет и отель.
    
    Используй следующий шаблон:
    
    Текст: сообщение пользователя в чате
    Класс: одно из двух значений: самолет или отель
    
    Начинай!
    
    Текст: {sample.text}
    Класс: """
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text
    generated_text = ''.join([char for char in generated_text if char.isalpha()])
    prediction.append(generated_text)

df['prediction'] = prediction
booking_errors = df[(df['prediction']!=df['label'])&(df['label']=='отель')].shape[0]
ticket_errors = df[(df['prediction']!=df['label'])&(df['label']=='самолет')].shape[0]
print(f'Accuracy booking = {1-booking_errors/1000:.3f}\nAccuracy tickets = {1-ticket_errors/1000:.3f}')

## One-shot learning

prediction_one_shot = []
sampling_params = SamplingParams(temperature=0, max_tokens=30, stop=['отель', 'самолет'], include_stop_str_in_output=True)
for i in tqdm(range(2000)):
    sample = df.iloc[i]
    prompt = f"""Ты должен отнести текст к одному из двух классов: самолет и отель.

    Пример для класса самолет: {df.iloc[0].text}
    Пример для класса отель: {df.iloc[1000].text}
    
    Используй следующий шаблон:
    
    Текст: сообщение пользователя в чате
    Класс: одно из двух значений: самолет или отель
    
    Начинай!
    
    Текст: {sample.text}
    Класс: """
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text
    generated_text = ''.join([char for char in generated_text if char.isalpha()])
    prediction_one_shot.append(generated_text)

df['prediction_one_shot'] = prediction_one_shot
booking_errors = df[(df['prediction_one_shot']!=df['label'])&(df['label']=='отель')].shape[0]
ticket_errors = df[(df['prediction_one_shot']!=df['label'])&(df['label']=='самолет')].shape[0]
print(f'Accuracy booking = {1-booking_errors/1000:.3f}\nAccuracy tickets = {1-ticket_errors/1000:.3f}')
