==> 3.NER.ipynb <==
# Тестирование in-context learning для Mistral-7B-Instruct-v0.2

from vllm import LLM, SamplingParams
from tqdm import tqdm

llm = LLM(model="mistralai/Mistral-7B-Instruct-v0.2")

import pandas as pd
import json

df = pd.read_csv('Data/generated_data.csv')
df.json = df.json.apply(lambda x: eval(x))

from utils import *
ner = NER()

sampling_params = SamplingParams(temperature=0, max_tokens=100, stop='}', include_stop_str_in_output=True)

for i in tqdm(range(2000)):
    sample = df.iloc[i]
    prompt = f"""Ты должен найти в тексте определенные атрибуты и сохранить их в формате JSON. Не используй комментарии.
    
    Используй следующий шаблон:
    
    Текст: сообщение пользователя в чате.
    JSON: объект по всем стандартам JSON.{ner.label_mapping[sample.label]['template']}
    Начинай!
    
    Текст: {sample.text}
    JSON:"""
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text
    ner.count_errors(sample, generated_text)

ner.score()

__Вывод:__
Модель с трудом определяет верные даты и числа, однако с названиями городов и отелей справляется хорошо.

__Дальнейшие шаги:__
Необходимо дообучить модель на задачу NER используя 80% данных. Предварительно нужно замерить точность на тестовой выборке, чтобы сравнить улучшение результата после тренировки.

# Точность на тестовых данных

sampling_params = SamplingParams(temperature=0, max_tokens=100, stop='}', include_stop_str_in_output=True)

test_ner = NER()

test_df = pd.concat([df.iloc[800:1000], df.iloc[1800:2000]])

for i in tqdm(range(400)):
    sample = test_df.iloc[i]
    prompt = f"""Ты должен найти в тексте определенные атрибуты и сохранить их в формате JSON. Не используй комментарии.

Используй следующий шаблон:

Текст: сообщение пользователя в чате.
JSON: объект по всем стандартам JSON.{test_ner.label_mapping[sample.label]['template']}
Начинай!

Текст: {sample.text}
Ответ:"""
    
    output = llm.generate(prompt, sampling_params, use_tqdm=False)[0]
        
    # Print the outputs.
    prompt = output.prompt
    generated_text = output.outputs[0].text
    test_ner.count_errors(sample, generated_text)

test_ner.score()
